[SPARK-9132][SPARK-9163][SQL] codegen conv
